refractoring of code
